+++ flask_praetorian/exceptions.py
@@ -1,53 +1,78 @@
-import flask
-import buzz
+import flask_buzz
 
 
-class PraetorianError(buzz.Buzz):
+class PraetorianError(flask_buzz.FlaskBuzz):
     """
-    Provides a custom exception class for flask-praetorian based on Buzz.
-    `py-buzz on gitub <https://github.com/dusktreader/py-buzz>`_
+    Provides a custom exception class for flask-praetorian based on flask-buzz.
+    `flask-buzz on gitub <https://github.com/dusktreader/flask-buzz>`_
     """
-
-    def __init__(self, *format_args, status_code=401, **format_kwds):
-        super().__init__(*format_args, **format_kwds)
-        self.status_code = status_code
-        self.headers = None
-
-    def jsonify(self):
-        """
-        Returns a representation of the error in a jsonic form that is
-        compatible with flask's error handling
-        """
-        return flask.jsonify({
-            'status_code': self.status_code,
-            'error': repr(self),
-            'description': self.message,
-        })
+    status_code = 401
 
 
 class MissingClaimError(PraetorianError):
+    """
+    The jwt token is missing a required claim
+    """
     pass
 
 
 class BlacklistedError(PraetorianError):
-    pass
+    """
+    The jwt token has been blacklisted and may not be used any more
+    """
+    status_code = 403
 
 
 class ExpiredAccessError(PraetorianError):
+    """
+    The jwt token has expired for access and must be refreshed
+    """
     pass
 
 
 class EarlyRefreshError(PraetorianError):
+    """
+    The jwt token has not yet expired for access and may not be refreshed
+    """
     pass
 
 
 class ExpiredRefreshError(PraetorianError):
+    """
+    The jwt token has expired for refresh. An entirely new token must be issued
+    """
     pass
 
 
 class MissingTokenHeader(PraetorianError):
+    """
+    The header is missing the required jwt token
+    """
     pass
 
 
 class InvalidTokenHeader(PraetorianError):
+    """
+    The token contained in the header is invalid
+    """
     pass
+
+
+class InvalidUserError(PraetorianError):
+    """
+    The user is no longer valid and is now not authorized
+    """
+    status_code = 403
+
+
+class MissingRoleError(PraetorianError):
+    """
+    The token is missing a required role
+    """
+    status_code = 403
+
+
+class MissingUserError(PraetorianError):
+    """
+    The user could not be identified
+    """